Move form validator to shared [ISDK-77]
